+++ src/main/java/dev/r1nex/cases/command/CaseCommand.java
@@ -103,7 +103,7 @@ public class CaseCommand extends AbstractCommand {
                         ItemStack itemStack = new ItemStack(Material.valueOf(section.getString(key + ".item")));
                         String displayName = section.getString(key + ".display-name");
                         List<String> action = section.getStringList(key + ".action");
-                        int chance = section.getInt(key + ".chance");
+                        double chance = section.getDouble(key + ".chance");
 
                         GroupData data = new GroupData(itemStack, displayName, chance, action);
                         groupData.add(data);
